Move client iterator to its own package
