+++ hazelcast/src/main/java/com/hazelcast/client/cache/impl/ClientCacheProxySupport.java
@@ -450,7 +450,7 @@ abstract class ClientCacheProxySupport<K, V> extends ClientProxy implements ICac
 
     protected boolean callReplaceSync(K key, Data keyData, V newValue, Data newValueData,
                                       Data oldValueData, Data expiryPolicyData) {
-        InternalCompletableFuture<Boolean> future = doReplaceOnServer(keyData, newValueData,
+        CompletableFuture<Boolean> future = doReplaceOnServer(keyData, newValueData,
                 oldValueData, expiryPolicyData, true, null);
         try {
             return future.get();
@@ -491,10 +491,10 @@ abstract class ClientCacheProxySupport<K, V> extends ClientProxy implements ICac
         return doReplaceOnServer(keyData, newValueData, oldValueData, expiryPolicyData, withCompletionEvent, statsCallback);
     }
 
-    private InternalCompletableFuture<Boolean> doReplaceOnServer(Data keyData, Data newValueData,
-                                                                 Data oldValueData, Data expiryPolicyData,
-                                                                 boolean withCompletionEvent,
-                                                                 BiConsumer<Boolean, Throwable> statsCallback) {
+    private CompletableFuture<Boolean> doReplaceOnServer(Data keyData, Data newValueData,
+                                                         Data oldValueData, Data expiryPolicyData,
+                                                         boolean withCompletionEvent,
+                                                         BiConsumer<Boolean, Throwable> statsCallback) {
         int completionId = withCompletionEvent ? nextCompletionId() : -1;
         ClientMessage request = CacheReplaceCodec.encodeRequest(nameWithPrefix, keyData, oldValueData, newValueData,
                 expiryPolicyData, completionId);
@@ -541,10 +541,10 @@ abstract class ClientCacheProxySupport<K, V> extends ClientProxy implements ICac
         }
     }
 
-    private <T> InternalCompletableFuture<T> doGetAndReplaceOnServer(Data keyData, Data newValueData,
-                                                                     Data expiryPolicyData,
-                                                                     boolean withCompletionEvent,
-                                                                     BiConsumer<T, Throwable> statsCallback) {
+    private <T> CompletableFuture<T> doGetAndReplaceOnServer(Data keyData, Data newValueData,
+                                                             Data expiryPolicyData,
+                                                             boolean withCompletionEvent,
+                                                             BiConsumer<T, Throwable> statsCallback) {
         int completionId = withCompletionEvent ? nextCompletionId() : -1;
         ClientMessage request = CacheGetAndReplaceCodec.encodeRequest(nameWithPrefix, keyData,
                 newValueData, expiryPolicyData, completionId);
@@ -829,7 +829,7 @@ abstract class ClientCacheProxySupport<K, V> extends ClientProxy implements ICac
         }
     }
 
-    protected InternalCompletableFuture<V> getAsyncInternal(Object key, ExpiryPolicy expiryPolicy) {
+    protected CompletableFuture<V> getAsyncInternal(Object key, ExpiryPolicy expiryPolicy) {
         long startNanos = nowInNanosOrDefault();
         BiConsumer<V, Throwable> statsCallback = statisticsEnabled
                 ? statsHandler.newOnGetCallback(startNanos) : null;
@@ -837,16 +837,16 @@ abstract class ClientCacheProxySupport<K, V> extends ClientProxy implements ICac
     }
 
     @Nonnull
-    protected InternalCompletableFuture<V> callGetAsync(Object key, ExpiryPolicy expiryPolicy,
-                                                        BiConsumer<V, Throwable> statsCallback) {
+    protected CompletableFuture<V> callGetAsync(Object key, ExpiryPolicy expiryPolicy,
+                                                BiConsumer<V, Throwable> statsCallback) {
         Data dataKey = toData(key);
         ClientDelegatingFuture<V> future = getInternal(dataKey, expiryPolicy, true);
-        addCallback(future, statsCallback);
-        return future;
+        return addCallback(future, statsCallback);
     }
 
-    protected void getAllInternal(Set<? extends K> keys, Collection<Data> dataKeys, ExpiryPolicy expiryPolicy,
-                                  List<Object> resultingKeyValuePairs, long startNanos) {
+    protected void getAllInternal(Set<? extends K> keys, Collection<Data> dataKeys,
+                                  ExpiryPolicy expiryPolicy, List<Object> resultingKeyValuePairs,
+                                  long startNanos) {
         if (dataKeys.isEmpty()) {
             objectToDataCollection(keys, dataKeys, getSerializationService(), NULL_KEY_IS_NOT_ALLOWED);
         }
@@ -870,7 +870,8 @@ abstract class ClientCacheProxySupport<K, V> extends ClientProxy implements ICac
         setExpiryPolicyInternal(keys, policy, null);
     }
 
-    protected void setExpiryPolicyInternal(Set<? extends K> keys, ExpiryPolicy policy, Set<Data> serializedKeys) {
+    protected void setExpiryPolicyInternal(Set<? extends K> keys, ExpiryPolicy policy,
+                                           Set<Data> serializedKeys) {
         try {
             List<Data>[] keysByPartition = groupKeysToPartitions(keys, serializedKeys);
             setExpiryPolicyAndWaitForCompletion(keysByPartition, policy);
